+++ frontend/src/components/screens/bookings/customers/assembly_terms/page.tsx
@@ -126,7 +126,7 @@ function AssemblyTermScreen() {
             {
                 loading ? (
                     <LoadingScreen />
-                ) : isLoggedIn && user?.user_role === "admin" ? (
+                ) : isLoggedIn && user?.user_role === "admin" || user?.user_role === "manager" ? (
                     <>
                         <Sidebar />
                         <main className='container p-1'>