+++ frontend/src/components/screens/bookings/staff/view_ticket/page.tsx
@@ -486,7 +486,7 @@ const ViewTicketStaffScreen = () => {
                                         <div className='flex items-center gap-4 md:justify-between my-2'>
                                             <h5 className="font-medium text-sm text-gray-500">Status</h5>
                                             <div className="relative">
-                                                {isLoggedIn && user?.user_role === "admin" ?
+                                                {isLoggedIn && user?.user_role === "admin" || user?.user_role === "manager" ?
                                                     <select name='unit_status' className="block w-full appearance-none rounded-md border border-gray-300 bg-white px-4 py-2 pr-8 text-sm shadow-sm focus:outline-none cursor-pointer [&>span]:line-clamp-1" value={unit_status || ''}
                                                         onChange={updateStatus}>
                                                         <option disabled value={""}>Select status</option>