+++ frontend/src/components/screens/cars/page.tsx
@@ -79,7 +79,7 @@ const CarsScreen = () => {
             {
                 loading ? (
                     <LoadingScreen />
-                ) : isLoggedIn && user?.user_role === "admin" ? (
+                ) : isLoggedIn && user?.user_role === "admin" || user?.user_role === "manager" ? (
                     <>
                         <Sidebar />
                         <main className='container p-1'>