+++ frontend/src/components/screens/driver_app/drivers/page.tsx
@@ -91,7 +91,7 @@ const DriversScreen = () => {
             {
                 loading ? (
                     <LoadingScreen />
-                ) : isLoggedIn && user?.user_role === "admin" ? (
+                ) : isLoggedIn && user?.user_role === "admin" || user?.user_role === "manager" ? (
                     <>
                         <Sidebar />
                         <main className='container p-1'>