+++ server/controllers/comments/add_comments.js
@@ -21,20 +21,16 @@ const addComment = async (req, res) => {
             "SELECT * FROM technician_tasks_comments WHERE task_id = $1 AND comment = $2 limit 1",
             [task_id, comment]
         );
-        // todo: remove
-        if (existingComment.length > 0) {
-            return res.status(500).json({ message: "comment exists" });
-        } else {
-            const { rows } = await pool.query(
-                "INSERT INTO technician_tasks_comments (task_id, comment, created_at, created_by) VALUES ($1, $2, $3, $4) returning *",
-                [task_id, comment, created_at, created_by]
-            );
-            await appLogs("INSERT", created_by, req.body, ticket_number);
-            return res.status(201).json({
-                message: "Successfully created",
-                rows: rows[0],
-            });
-        }
+
+        const { rows } = await pool.query(
+            "INSERT INTO technician_tasks_comments (task_id, comment, created_at, created_by) VALUES ($1, $2, $3, $4) returning *",
+            [task_id, comment, created_at, created_by]
+        );
+        await appLogs("INSERT", created_by, req.body, ticket_number);
+        return res.status(201).json({
+            message: "Successfully created",
+            rows: rows[0],
+        });
     } catch (error) {
         console.log("error add comment", error);
         // Handle validation or other errors