+++ server/middleware/verify_admin.js
@@ -35,28 +35,3 @@ export function authenticateAdmin(req, res, next) {
     req.user = user; // Attach user object to the request
     next();
 }
-export function authenticateRole(allowedRoles) {
-    return (req, res, next) => {
-        try {
-            const token = req.cookies.accessToken; // Use accessToken instead of refreshToken
-            if (!token)
-                return res
-                    .status(401)
-                    .json({ error: "Authentication required" });
-
-            const user = verifyJwtToken(token);
-            if (!user)
-                return res
-                    .status(403)
-                    .json({ error: "Invalid or expired token" });
-
-            if (!allowedRoles.includes(user.user_role))
-                return res.status(403).json({ error: "Access denied." });
-
-            req.user = user;
-            next();
-        } catch (error) {
-            return res.status(500).json({ error: "Internal server error" });
-        }
-    };
-}