+++ server/routes/department/hhp/hhp_jobs_route.js
@@ -14,6 +14,7 @@ import { limiter } from "../../../middleware/rateLimiter.js";
 import { authenticateToken } from "../../../middleware/verify.js";
 import { authenticateAdmin } from "../../../middleware/verify_admin.js";
 import AddToQCTable from "../../../controllers/department/hhp/technicians/add_to_qc_table.js";
+import { authenticateRole } from "../../../middleware/verify_role.js";
 const router = express.Router();
 
 router.post("/", limiter, authenticateToken, AddHHPTask);
@@ -21,10 +22,14 @@ router.patch("/assess/:id", limiter, authenticateToken, UpdateAssessmentDate);
 router.post("/qc_table", AddToQCTable);
 router.get("/", authenticateToken, GetAllTasks);
 router.get("/:id", authenticateToken, GetTaskById);
-router.get("/engineer/bin", authenticateAdmin, GetBinStats);
+router.get(
+    "/engineer/bin",
+    authenticateRole(["admin", "manager"]),
+    GetBinStats
+);
 router.get("/ticket/:id", limiter, GetTaskByTicket);
 router.patch("/:id", authenticateToken, UpdateTask);
 router.patch("/so/:id", authenticateToken, UpdateSOAfterBooking);
-router.delete("/:id", authenticateAdmin, deleteHHPTask);
+router.delete("/:id", authenticateRole(["manager"]), deleteHHPTask);
 
 export { router };